pokeapi/requirements.txt
django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0
djoser==2.1.0
psycopg2==2.8